+++ src/components/Filter/Context/FilterReducer.ts
@@ -1,41 +1,8 @@
-import { ActionType, createCustomAction, getType } from 'typesafe-actions';
-import { FilterData } from '../Types/FilterItem';
+import { getType } from 'typesafe-actions';
+import { actions, Actions } from './FilterActions';
 import { FilterState } from './FilterContext';
 
-export enum FilterActions {
-    setFilteredData = 'setFilteredData',
-    setFilter = 'setFilter',
-    setData = 'setData',
-    setIsLoading = 'setIsLoading'
-}
-
-export const actions = {
-    setFilter: createCustomAction(
-        FilterActions.setFilter,
-        (filter: FilterData) => ({
-            filter
-        })
-    ),
-    setData: createCustomAction(FilterActions.setData, (data) => ({
-        data
-    })),
-    setFilteredData: createCustomAction(
-        FilterActions.setFilteredData,
-        (filteredData) => ({
-            filteredData
-        })
-    ),
-    setIsLoading: createCustomAction(
-        FilterActions.setIsLoading,
-        (isLoading: boolean) => ({
-            isLoading
-        })
-    )
-};
-
-export type Action = ActionType<typeof actions>;
-
-export function filterReducer(state: FilterState, action: Action): FilterState {
+export function filterReducer(state: FilterState, action: Actions): FilterState {
     switch (action.type) {
         case getType(actions.setData):
             return { ...state, data: action.data };
